extend the GAP library functions `GO`, `GU`, etc.
... such that one can prescribe an invariant form.
This will work as soon as the corresponding extensions
in the GAP library become available.

Some tests of the new code are provided in `tst/classic.tst`;
currently they are not processed automatically when `tst/testall.g` is run,
note that they require the corresponding changes in the GAP library.

(Documentation for the new functions is missing.)