Use a test for GAP 4.12 that works in CI, too
The problem is that we don't use actual old GAP releases in CI,
but rather GAP builds made from the corresponding stable-4.X branches.
In there, the GAP version is something like 4.dev, which is always
treated as "newer" than any other version. So we have to be a bit
creative to make it work.